backend/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
boto3==1.28.55
botocore==1.31.55
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
comm==0.1.4
cryptography==41.0.3
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
django-storages==1.14
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
exceptiongroup==1.1.3
executing==1.2.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jmespath==1.0.1
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.7.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.3
traitlets==5.10.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
zipp==3.17.0
